Removing basic_auth copy.
